requirements.txt
pytest
requests
pytest-html
python-dotenv

